+++ src/main/java/com/iexec/worker/config/CoreConfigurationService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,11 @@
 
 package com.iexec.worker.config;
 
+import com.iexec.core.api.SchedulerClient;
+import com.iexec.core.api.SchedulerClientBuilder;
+import feign.Logger;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
@@ -66,5 +70,10 @@ public class CoreConfigurationService {
 
     public void setCoreSessionId(String coreSessionId) {
         this.coreSessionId = coreSessionId;
+    }
+
+    @Bean
+    SchedulerClient schedulerClient() {
+        return SchedulerClientBuilder.getInstance(Logger.Level.FULL, url.toString());
     }
 }